完成岗位及组织机构树功能开发

+++ src/main/java/com/cg/cm/organize/controller/ZCMT6001Controller.java
@@ -3,11 +3,12 @@ package com.cg.cm.organize.controller;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.cg.cm.organize.entity.ZCMT6001;
+import com.cg.cm.organize.entity.ZCMT6002;
 import com.cg.cm.organize.service.ZCMT6001Service;
+import com.cg.cm.organize.service.ZCMT6002Service;
 import com.demo.cm.utils.BaseResponse;
 import com.demo.cm.utils.CMException;
 import com.google.common.base.Strings;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.web.bind.annotation.*;
@@ -23,15 +24,20 @@ import java.util.List;
 
 @RestController
 public class ZCMT6001Controller {
-    @Autowired
     private ZCMT6001Service service;
+    private ZCMT6002Service zcm6002service;
     private BaseResponse resp;
+
+    public ZCMT6001Controller(ZCMT6001Service service, ZCMT6002Service zcm6002service) {
+        this.service = service;
+        this.zcm6002service = zcm6002service;
+    }
 
     /**
      * 增加部门（不用）
      */
-    @RequestMapping(value = "/department/addbdepart" ,method = RequestMethod.POST)
-    public String addDdepart(@RequestBody ZCMT6001 zcmt6001){
+    @RequestMapping(value = "/department/adddepart" ,method = RequestMethod.POST)
+    public String addDepart(@RequestBody ZCMT6001 zcmt6001){
         resp = new BaseResponse();
         if(Strings.isNullOrEmpty(zcmt6001.getBukrs())){
             return resp.setStatecode(BaseResponse.ERROR).setMsg("公司代码不能为空！").toJSON();
@@ -40,7 +46,7 @@ public class ZCMT6001Controller {
             return resp.setStatecode(BaseResponse.ERROR).setMsg("部门代码不能为空！").toJSON();
         }
         try{
-            if(service.getDPNAM(zcmt6001.getDpnum()) != null){
+            if(service.getDPNUM(zcmt6001.getDpnum()) != null){
                 return resp.setStatecode(BaseResponse.ERROR).setMsg("部门代码已存在").toJSON();
             }
             zcmt6001.setCreateDate(Calendar.getInstance().getTime());
@@ -110,11 +116,12 @@ public class ZCMT6001Controller {
             if( zcmt6001== null){
                 return resp.setStatecode(BaseResponse.ERROR).setMsg("部门代码不存在！").toJSON();
             }
-            // TODO 挂有岗位的部门不可删除
-//            int size = service.getChild(bukrs.getBukrs()).size();
-//            if(size>0){
-//                return resp.setStatecode(BaseResponse.ERROR).setMsg("该公司有个"+size+"子公司，不可删除！").toJSON();
-//            }
+            // 挂有岗位的部门不可删除
+            List<ZCMT6002> zcmt6002 = zcm6002service.getAll(dpnum.getDpnum());
+            if(zcmt6002 != null && zcmt6002.size()>0){
+                return resp.setStatecode(BaseResponse.ERROR).setMsg("该部门下有"+zcmt6002.size()+"个岗位，不可删除！").toJSON();
+            }
+            // 删除部门
             service.deleteZCMT6001(zcmt6001);
             if(service.getDPNUM(dpnum.getDpnum()) == null){
                 return resp.setStatecode(BaseResponse.SUCCESS).setMsg("删除成功！").toJSON();
@@ -138,7 +145,7 @@ public class ZCMT6001Controller {
         try{
             List<ZCMT6001> list = service.getAll(zcmt6001.getBukrs());
             if( list== null){
-                return resp.setStatecode(BaseResponse.ERROR).setMsg("部门为空！").toJSON();
+                return resp.setStatecode(BaseResponse.EMPTY).setMsg("部门为空！").toJSON();
             }
             return resp.setStatecode(BaseResponse.SUCCESS).setData(list).toJSON();
         }catch (CMException e){
@@ -159,6 +166,9 @@ public class ZCMT6001Controller {
         try{
             Page<ZCMT6001> list = service.pageAll(jsonObject.getString("bukrs"),
                     new PageRequest(jsonObject.getIntValue("page"),jsonObject.getIntValue("size")));
+            if( list== null){
+                return resp.setStatecode(BaseResponse.EMPTY).setMsg("部门为空！").toJSON();
+            }
             return resp.setStatecode(BaseResponse.SUCCESS).setData(list).toJSON();
         }catch (CMException e){
             return resp.setStatecode(BaseResponse.ERROR).setMsg(e.getMsg()).toJSON();
